New module kas-core::core::events
